test: Separate account test
